added useMemo
