+++ .eslintrc.json
@@ -15,6 +15,8 @@
     "class-methods-use-this": "off",
     "import/prefer-default-export": "off",
     "react/react-in-jsx-scope": "off",
-    "react/function-component-definition": "off"
+    "react/function-component-definition": "off",
+    "jsx-a11y/click-events-have-key-events": "off",
+    "jsx-a11y/no-static-element-interactions": "off"
   }
 }
