+++ src/components/features/sidebar.tsx
@@ -1,17 +1,4 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import { useMemo } from 'react';
-import { FiNavigation } from 'react-icons/fi';
-
-export const cityLocations = [
-  { city: 'Москва', branch: 92 },
-  { city: 'Санкт-Петербург', branch: 16 },
-  { city: 'Калининград', branch: 10 },
-  { city: 'Красногорск', branch: 5 },
-  { city: 'Воронеж', branch: 4 },
-  { city: 'Долгопрудный', branch: 3 },
-  { city: 'Красноярск', branch: 3 },
-  { city: 'Раменское', branch: 3 },
-];
+import { Location } from './components/cities';
 
 interface ModalWindow {
   isOpen: boolean;
@@ -81,29 +68,9 @@ export function Sidebar({ isOpen, setIsOpen }: ModalWindow): JSX.Element {
                   </div>
                 </form>
 
-                {/* Hardcoded cities */}
+                {/* Hardcoded cities imported */}
                 <div>
-                  {useMemo(
-                    () =>
-                      cityLocations.map((el) => (
-                        <>
-                          <div className="flex p-2 m-2 font-button_font">
-                            <FiNavigation
-                              color="#BFA15C"
-                              size={20}
-                              fill="#BFA15C"
-                              className="mt-4"
-                            />
-                            <div className="ml-4">
-                              <p className="font-medium text-gold-title">{el.city}</p>
-                              <p>{el.branch} филиалов</p>
-                            </div>
-                          </div>
-                          <hr className="bg-black h-0.5 w-11/12 ml-4" />
-                        </>
-                      )),
-                    [],
-                  )}
+                  <Location />
                 </div>
               </div>
             </div>
